类型改变

+++ cloud-common/src/main/java/com/cloud/common/enums/OrderImeiStatus.java
@@ -6,22 +6,22 @@ package com.cloud.common.enums;
  */
 public enum OrderImeiStatus {
     /** 发货前提供串码 */
-    PRE_DELIVERY(1, "发货前提供串码"),
+    PRE_DELIVERY(1L, "发货前提供串码"),
     /** 发货后提供串码 */
-    AFTER_DELIVERY(2, "发货后提供串码"),
+    AFTER_DELIVERY(2L, "发货后提供串码"),
     /** 不需要提供串码 */
-    NO_IMEI(3, "不需要提供串码");
+    NO_IMEI(3L, "不需要提供串码");
 
-    private final Integer code;
+    private final Long code;
     private final String msg;
 
 
-    OrderImeiStatus(Integer code, String msg) {
+    OrderImeiStatus(Long code, String msg) {
         this.code = code;
         this.msg = msg;
     }
 
-    public Integer getCode() {
+    public Long getCode() {
         return code;
     }
